feat: allow settings any year other than only above 1970
We don't have constraint in `toParts`, thus we can still have a year lower than 1970.

One way to make `withYear` and `toParts` match is to modify `toParts` to add the same constraint about year. But that
means the parts we get from `toParts` doesn't correspond to the original value. Thus we decide to allow settings any
year in `withYear`